+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Rank/epoch typo in perf chart tooltip (#408)
+-   Add channel namespace to values stored in localStorage (#412)
 -   Catch cancel error when voluntarily aborting CP details calls (#414)
 
 ### Changed